pyproject.toml
[project]
name = "hexlet-code"
version = "0.1.0"
description = "diff searcher and formatter of result"
authors = [{name = "Alsy", email="velsych@mail.ru"}]
classifiers = [
    "Environment :: Console"]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pyyaml>=6.0.2",
]


[project.scripts]
gendiff = "gendiff.scripts.gendiff:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["gendiff"]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "pytest-cov>=6.2.1",
    "ruff>=0.12.8",
]
